requirements.txt
gspread
google-auth
python-dotenv
google-genai
openai

